# internal/deephash: hash maps without sorting in the acyclic common case