@bradfitz 2021-05-14
This is blocked on https://github.com/golang/go/issues/46131 (which we can do in our own Go fork first).