@josharian 2021-05-20
OK, I'm finally done tinkering with this. Results for the whole PR, when built with `-tags=tailscale_go`:

```
name    old time/op    new time/op    delta
Hash-8    21.2µs ± 0%    12.5µs ± 0%  -40.83%  (p=0.008 n=5+5)

name    old alloc/op   new alloc/op   delta
Hash-8    2.88kB ± 0%    0.79kB ± 0%  -72.47%  (p=0.008 n=5+5)

name    old allocs/op  new allocs/op  delta
Hash-8      86.0 ± 0%      12.0 ± 0%  -86.05%  (p=0.008 n=5+5)
```

Most importantly, most of the remaining allocs are per call to UpdateHash, not proportional to the input to UpdateHash.

One exception to that is that we allocate one iteration per map. Fixing that would require new API; I've filed https://github.com/golang/go/issues/46293.

The other exception to that is when the per-type optimizations are working with non-addressable values. In that case, we allocate once per value. I've fixed the only case in which that happens in the existing code, wgengine/userspace.go. If it matters, we could use reflect.Value.Set in conjunction with more scratch values or use more sync.Pools. But this is good enough for now.

This branch could probably use re-working, re-ordering, squashing, etc., but I'm kind of worn out on it. We can sort out what to do next week when you're around to review, Brad.